move CDI Config test files and utility functions
